refactor: convert JS files to TS and update syntax for TypeScript

+++ backend/src/controllers/events.controllers.ts
@@ -1,6 +1,7 @@
 import textModel from "../models/Text.js";
+import { RequestHandler } from "express";
 
-export const addText = async (req, res) => {
+export const addText: RequestHandler = async (req, res) => {
   try {
     const userId = req.user.userId;
     const { text } = req.body;
@@ -10,21 +11,24 @@ export const addText = async (req, res) => {
       text,
     });
     await newText.save();
-    return res.status(200).json({ text: newText.text, status: newText.status });
-  } catch {
-    (error) => {
-      res.status(500).json("Server Error: ", error.message);
-    };
+    return res
+      .status(200)
+      .json({ text: newText.text, isCompleted: newText.isCompleted });
+  } catch (error) {
+    if (error instanceof Error) {
+      console.log("Error: ", error);
+      res.status(500).json({ msg: "Server Error", error: error.message });
+    }
   }
 };
 
-export const getAllText = async (req, res) => {
+export const getAllText: RequestHandler = async (req, res) => {
   try {
     const sort = req.body.sort;
     const userId = req.user.userId;
     const userName = req.user.userName;
     console.log("GET ALL TEXT: ", userId, sort);
-    let userData = [];
+    let userData: any[] = [];
     if (sort === "all" || sort === "") {
       userData = await textModel.find({ createdBy: userId });
     } else if (sort === "pending") {
@@ -35,18 +39,20 @@ export const getAllText = async (req, res) => {
     } else if (sort === "completed") {
       userData = await textModel.find({ createdBy: userId, isCompleted: true });
     }
-
-    if (!userData.length === 0) {
+    if (userData.length === 0) {
       return res.status(404).json({ msg: "No Data Found" });
     } else {
       return res.status(200).json({ userData: userData, userName: userName });
     }
   } catch (error) {
-    return res.status(500).json({ error: error.message });
+    if (error instanceof Error) {
+      console.log("Error: ", error);
+      res.status(500).json({ msg: "Server Error", error: error.message });
+    }
   }
 };
 
-export const deleteById = async (req, res) => {
+export const deleteById: RequestHandler = async (req, res) => {
   try {
     const textId = req.params.id;
     console.log("DELETE BY ID: ", textId);
@@ -54,6 +60,9 @@ export const deleteById = async (req, res) => {
       return res.status(404).json({ msg: "Id not found." });
     } else {
       const userData = await textModel.findById(textId);
+      if (!userData) {
+        return res.json({ msg: "No user data found" });
+      }
       const deleted = await textModel.findByIdAndDelete(textId);
       if (deleted) {
         return res
@@ -62,11 +71,14 @@ export const deleteById = async (req, res) => {
       }
     }
   } catch (error) {
-    console.error(error.message);
+    if (error instanceof Error) {
+      console.log("Error: ", error);
+      res.status(500).json({ msg: "Server Error", error: error.message });
+    }
   }
 };
 
-export const editById = async (req, res) => {
+export const editById: RequestHandler = async (req, res) => {
   try {
     const userId = req.params.id;
     const { text } = req.body;
@@ -74,20 +86,28 @@ export const editById = async (req, res) => {
     if (!userId) {
       return res.status(404).json({ msg: "Id not found." });
     } else {
-      const done = await textModel.findByIdAndUpdate(userId, { text: text });
-      const updated = await textModel.findById(userId);
+      const done = await textModel.findByIdAndUpdate(
+        userId,
+        { text: text },
+        { new: true },
+      );
       if (done) {
         return res
           .status(200)
-          .json({ msg: "Update Successful", updatedData: updated.text });
+          .json({ msg: "Update Successful", updatedData: done.text });
+      } else {
+        return res.status(500).json({ msg: "Updating error" });
       }
     }
   } catch (error) {
-    console.error(error.message);
+    if (error instanceof Error) {
+      console.log("Error: ", error);
+      res.status(500).json({ msg: "Server Error", error: error.message });
+    }
   }
 };
 
-export const status = async (req, res) => {
+export const status: RequestHandler = async (req, res) => {
   try {
     const userId = req.params.id;
     const userStatus = req.body.isCompleted;
@@ -95,17 +115,25 @@ export const status = async (req, res) => {
     if (!userId) {
       return res.status(404).json({ msg: "Id not found." });
     } else {
-      const done = await textModel.findByIdAndUpdate(userId, {
-        isCompleted: userStatus,
-      });
-      const updated = await textModel.findById(userId);
+      const done = await textModel.findByIdAndUpdate(
+        userId,
+        {
+          isCompleted: userStatus,
+        },
+        { new: true },
+      );
       if (done) {
         return res
           .status(200)
-          .json({ msg: "Status Changed.", status: updated.isCompleted });
+          .json({ msg: "Status Changed.", status: done.isCompleted });
+      } else {
+        return res.status(500).json({ msg: "Updating error" });
       }
     }
   } catch (error) {
-    console.error(error.message);
+    if (error instanceof Error) {
+      console.log("Error: ", error);
+      res.status(500).json({ msg: "Server Error", error: error.message });
+    }
   }
 };
